requirements.txt
pillow
shutil
zipfile
time
argparse
subprocess